backend/requirements.txt
pypdf
langchain-community
langchain-huggingface
fastapi
uvicorn
python-dotenv
pydantic-settings
groq
chromadb
sentence-transformers
python-multipart
requests
pypdf
langchain-community
langchain-huggingface
rank-bm25
langchain-chroma